Codes/Web Application/Backend/requirements.txt
gunicorn
click
Flask
itsdangerous
Jinja2
MarkupSafe
python-dateutil
pytz
scikit-learn
scipy
six
threadpoolctl
Werkzeug
pandas
tensorflow
keras
pillow
numpy
opencv-python

